chronosscan/ai-service/requirements.txt
flask
flask-cors
numpy
opencv-python
torch
torchvision
sentence-transformers
faiss-cpu
scikit-learn
pymysql
python-dotenv
openai
requests
Pillow
scipy
